+++ sitegroup/src/main/java/com/zn/sitegroup/repository/IProductInfoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.Repository;
 /**
  * Created by zn on 2018/12/14.
  */
-public interface IProductInfoRepository extends JpaRepository<LcProductsInfoEntity,Integer> {
-    LcProductsInfoEntity findByProductId(int productId);
+public interface IProductInfoRepository extends JpaRepository<LcProductsInfoEntity,Long> {
+    LcProductsInfoEntity findByProductId(long productId);
 
 }
