+++ sitegroup/src/main/java/com/zn/sitegroup/repository/IProductToCategoryRepository.java
@@ -7,18 +7,18 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by zn on 2018/12/15.
  */
-public interface IProductToCategoryRepository extends JpaRepository<LcProductsToCategoriesEntity,Integer> {
+public interface IProductToCategoryRepository extends JpaRepository<LcProductsToCategoriesEntity,Long> {
     /**
      * 查找分类下有哪些商品
      * @param categoryId
      * @return
      */
-    List<LcProductsToCategoriesEntity> findProductsByCategoryId(int categoryId);
+    List<LcProductsToCategoriesEntity> findProductsByCategoryId(long categoryId);
 
     /**
      * 查找产品下有哪些分类
      * @param productId
      * @return
      */
-    List<LcProductsToCategoriesEntity> findByProductId(int productId);
+    List<LcProductsToCategoriesEntity> findByProductId(long productId);
 }
